Create a queue in RabbitMQ.

This will first check to see if the queue already exists in RabbitMQ, if not it will create the queue. If the queue exists, or is created it will return TRUE, it will return FALSE otherwise.

Usage

rabbitmq_create_queue(
  url,
  auth,
  vhost,
  queue,
  auto_delete = FALSE,
  durable = TRUE
)

Arguments

url

parsed RabbitMQ URL.

auth

the httr authentication object to use.

vhost

the vhost where to create the queue.

queue

the queue that should be checked/created.

auto_delete

should the queue be deleted afterwards (FALSE is default)

durable

should the messages exists after a server restart (TRUE is default)

Value

TRUE if the queue now exists, FALSE otherwise.

Author

Rob Kooper