Get message from RabbitMQ.

This will get a message from RabbitMQ, if the queue does not exist it will be created. The message will be converted to a json message that is returned.

Usage

rabbitmq_get_message(uri, queue, count = 1, prefix = "", port = 15672)

Arguments

uri

RabbitMQ URI or URL to rest endpoint

queue

the queue the message is received from.

count

the number of messages to retrieve from the queue.

prefix

prefix for the rabbitmq api endpoint, default is for no prefix.

port

port for the management interface, the default is 15672.

Value

NA if no message was retrieved, or a list of the messages payload.

Author

Alexey Shiklomanov, Rob Kooper